requirements.txt
langchain
qdrant-client
openai
tiktoken
python-dotenv
fastapi
uvicorn
schedule
langchain-qdrant